# Async optimization in scan path